RF_FooterTest_testFooterRestApiClickOnXmlApiDisplayXML -refactoring test add RestApiPage - FooterPage is abstract
