+++ src/test/java/model/HomePage.java
@@ -65,9 +65,6 @@ public class HomePage extends Header {
     @FindBy(xpath = "//div[@class='tabBar']/div/a")
     private List<WebElement> viewList;
 
-    @FindBy(xpath = "//a[@href='api/']")
-    private WebElement restApiLink;
-
     @FindBy(xpath = "//div/a[@class='model-link']")
     private WebElement iconUserName;
 
@@ -283,12 +280,6 @@ public class HomePage extends Header {
     public String getJobBuildStatus(String name) {
         return getDriver().findElement(By.id(String.format("job_%s", name)))
                 .findElement(By.xpath(".//*[name()='svg']")).getAttribute("tooltip");
-    }
-
-    public FooterPage clickRestApiLink() {
-        restApiLink.click();
-
-        return new FooterPage(getDriver());
     }
 
     public StatusUserPage clickUserIcon() {
