+++ src/test/java/model/XmlPage.java
@@ -4,16 +4,14 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class XmlPage extends BasePage {
+public class XmlPage extends RestApiPage {
     @FindBy(css = "body > div.header>span")
     private WebElement textOnPageXML;
 
     public XmlPage(WebDriver driver) {
         super(driver);
     }
-
     public String getStructureXML() {
-
         return textOnPageXML.getText();
     }
 }
